Fix odf's sqlx feature
