+++ src/domain/accounts/domain/Cargo.toml
@@ -21,12 +21,15 @@ workspace = true
 doctest = false
 
 
+[features]
+default = []
+sqlx = ["dep:sqlx"]
+
+
 [dependencies]
 database-common = { workspace = true }
 internal-error = { workspace = true }
-opendatafabric = { workspace = true, default-features = true, features = [
-    "sqlx",
-] }
+opendatafabric = { workspace = true }
 random-names = { workspace = true }
 
 async-trait = { version = "0.1", default-features = false }
@@ -41,9 +44,13 @@ rand = "0.8"
 reusable = "0.1"
 serde = "1"
 serde_with = { version = "3", default-features = false }
-sqlx = { version = "0.7", default-features = false, features = ["macros"] }
 thiserror = { version = "1", default-features = false }
 tracing = { version = "0.1", default-features = false }
 uuid = { version = "1", default-features = false, features = ["v4"] }
+
+# Optional
+sqlx = { optional = true, version = "0.7", default-features = false, features = [
+    "macros",
+] }
 
 [dev-dependencies]
