+++ src/domain/datasets/domain/Cargo.toml
@@ -21,24 +21,26 @@ workspace = true
 doctest = false
 
 
+[features]
+default = []
+sqlx = ["dep:sqlx"]
+
+
 [dependencies]
 database-common = { workspace = true }
 internal-error = { workspace = true }
-opendatafabric = { workspace = true, default-features = true, features = [
-    "sqlx",
-] }
+opendatafabric = { workspace = true }
 
 aes-gcm = { version = "0.10.3" }
 async-trait = { version = "0.1", default-features = false }
 chrono = { version = "0.4", default-features = false }
 secrecy = "0.8"
-sqlx = { version = "0.7", default-features = false, features = [
-    "macros",
-    "mysql",
-    "sqlite",
-    "postgres",
-] }
 thiserror = { version = "1", default-features = false }
 uuid = { version = "1", default-features = false, features = ["v4"] }
+
+# Optional
+sqlx = { optional = true, version = "0.7", default-features = false, features = [
+    "macros",
+] }
 
 [dev-dependencies]
