+++ src/domain/opendatafabric/src/identity/account_id.rs
@@ -10,7 +10,6 @@
 use ed25519_dalek::SigningKey;
 
 use crate::formats::*;
-use crate::impl_sqlx;
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
@@ -184,6 +183,6 @@ impl<'de> serde::de::Visitor<'de> for AccountIDSerdeVisitor {
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 #[cfg(feature = "sqlx")]
-impl_sqlx!(AccountID);
+super::sqlx::impl_sqlx!(AccountID);
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
