+++ src/domain/opendatafabric/src/identity/dataset_id.rs
@@ -11,7 +11,6 @@ use ed25519_dalek::SigningKey;
 
 use super::{DatasetRef, DatasetRefAny, DatasetRefRemote};
 use crate::formats::*;
-use crate::impl_sqlx;
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
@@ -209,6 +208,6 @@ impl<'de> serde::de::Visitor<'de> for DatasetIDSerdeVisitor {
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 #[cfg(feature = "sqlx")]
-impl_sqlx!(DatasetID);
+super::sqlx::impl_sqlx!(DatasetID);
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
