+++ src/domain/opendatafabric/src/identity/mod.rs
@@ -20,3 +20,6 @@ pub use dataset_id::*;
 pub use dataset_identity::*;
 pub use dataset_refs::*;
 pub use grammar::*;
+
+#[cfg(feature = "sqlx")]
+mod sqlx;
