+++ src/infra/accounts/postgres/Cargo.toml
@@ -23,9 +23,9 @@ doctest = false
 
 [dependencies]
 database-common = { workspace = true }
-kamu-accounts = { workspace = true }
 internal-error = { workspace = true }
-opendatafabric = { workspace = true }
+kamu-accounts = { workspace = true, features = ["sqlx"] }
+opendatafabric = { workspace = true, features = ["sqlx-postgres"] }
 
 async-trait = { version = "0.1", default-features = false }
 chrono = { version = "0.4", default-features = false }
@@ -34,7 +34,7 @@ sqlx = { version = "0.7", default-features = false, features = [
     "runtime-tokio-rustls",
     "macros",
     "postgres",
-    "chrono"
+    "chrono",
 ] }
 thiserror = { version = "1", default-features = false }
 tracing = { version = "0.1", default-features = false }
