+++ src/infra/datasets/postgres/Cargo.toml
@@ -23,9 +23,9 @@ doctest = false
 
 [dependencies]
 database-common = { workspace = true }
-kamu-datasets = { workspace = true }
+kamu-datasets = { workspace = true, features = ["sqlx"] }
 internal-error = { workspace = true }
-opendatafabric = { workspace = true }
+opendatafabric = { workspace = true, features = ["sqlx-postgres"] }
 
 async-trait = { version = "0.1", default-features = false }
 chrono = { version = "0.4", default-features = false }
@@ -35,7 +35,7 @@ sqlx = { version = "0.7", default-features = false, features = [
     "runtime-tokio-rustls",
     "macros",
     "postgres",
-    "chrono"
+    "chrono",
 ] }
 thiserror = { version = "1", default-features = false }
 tracing = { version = "0.1", default-features = false }
